Move chat linked list to shared project
