+++ SSW.Rules.GPT.sln
@@ -31,6 +31,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application.UnitTests", "te
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application.IntegrationTests", "tests\Application.IntegrationTests\Application.IntegrationTests.csproj", "{8F5EA9E9-6B28-4308-B8D0-A0C28E05204D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shared", "src\Shared\Shared.csproj", "{F86915AF-37EA-4B49-987A-3AE0978D1EBA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -65,6 +67,10 @@ Global
 		{8F5EA9E9-6B28-4308-B8D0-A0C28E05204D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8F5EA9E9-6B28-4308-B8D0-A0C28E05204D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8F5EA9E9-6B28-4308-B8D0-A0C28E05204D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F86915AF-37EA-4B49-987A-3AE0978D1EBA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F86915AF-37EA-4B49-987A-3AE0978D1EBA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F86915AF-37EA-4B49-987A-3AE0978D1EBA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F86915AF-37EA-4B49-987A-3AE0978D1EBA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -80,5 +86,6 @@ Global
 		{ECCD1CBB-0AE9-46CA-8B49-5A94F956F1E5} = {DF971C14-AD50-4B07-ACED-087B6ECA7500}
 		{BAE42F60-9354-4321-9A9A-A3F88D2C7236} = {D23E8309-9E03-4769-B6CC-2959CB36C8F5}
 		{8F5EA9E9-6B28-4308-B8D0-A0C28E05204D} = {D23E8309-9E03-4769-B6CC-2959CB36C8F5}
+		{F86915AF-37EA-4B49-987A-3AE0978D1EBA} = {DF971C14-AD50-4B07-ACED-087B6ECA7500}
 	EndGlobalSection
 EndGlobal
